requirements.txt
neo4j==5.15.0
py2neo==2021.2.4
spacy==3.7.2
scispacy==0.5.3
transformers==4.36.0
torch==2.1.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.0
numpy==1.26.0
pandas==2.1.0
scikit-learn==1.3.0
requests==2.31.0
beautifulsoup4==4.12.0
pdfplumber==0.10.0
python-multipart==0.0.6
python-dotenv==1.0.0
pytest==7.4.0
jupyter==1.0.0
networkx==3.2.0
matplotlib==3.8.0
seaborn==0.13.0
